Namespace function to_json_pretty()
